Home component now manages profile state, moved to its own file
